+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/model/entity/Favorite.kt
@@ -2,9 +2,9 @@ package com.sw.sw_api_kotlin_project.model.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.sw.sw_api_kotlin_project.network.model.Film
-import com.sw.sw_api_kotlin_project.network.model.People
-import com.sw.sw_api_kotlin_project.network.model.Planet
+import com.sw.sw_api_kotlin_project.data.network.model.Film
+import com.sw.sw_api_kotlin_project.data.network.model.People
+import com.sw.sw_api_kotlin_project.data.network.model.Planet
 
 @Entity(tableName = "favorite_table")
 data class Favorite(
