+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/model/repository/PeopleRepository.kt
@@ -3,8 +3,8 @@ package com.sw.sw_api_kotlin_project.model.repository
 import com.sw.sw_api_kotlin_project.model.data.PeoplePagingSource
 import com.sw.sw_api_kotlin_project.model.entity.Resource
 import com.sw.sw_api_kotlin_project.network.SWService
-import com.sw.sw_api_kotlin_project.network.model.People
-import com.sw.sw_api_kotlin_project.network.model.Results
+import com.sw.sw_api_kotlin_project.data.network.model.People
+import com.sw.sw_api_kotlin_project.data.network.model.Results
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
