+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/base/FilmNavListener.kt
@@ -1,6 +1,6 @@
 package com.sw.sw_api_kotlin_project.screen.base
 
-import com.sw.sw_api_kotlin_project.network.model.Film
+import com.sw.sw_api_kotlin_project.data.network.model.Film
 
 interface FilmNavListener {
     fun getFilmValue(): Film
