+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/base/PeopleNavListener.kt
@@ -1,6 +1,6 @@
 package com.sw.sw_api_kotlin_project.screen.base
 
-import com.sw.sw_api_kotlin_project.network.model.People
+import com.sw.sw_api_kotlin_project.data.network.model.People
 
 interface PeopleNavListener {
     fun getPeopleValue(): People
