+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/base/PlanetNavListener.kt
@@ -1,6 +1,6 @@
 package com.sw.sw_api_kotlin_project.screen.base
 
-import com.sw.sw_api_kotlin_project.network.model.Planet
+import com.sw.sw_api_kotlin_project.data.network.model.Planet
 
 interface PlanetNavListener {
     fun getPlanetValue(): Planet
