+++ SW_API_Kotlin_Project/app/src/main/res/navigation/nav_main_graph.xml
@@ -38,7 +38,7 @@
         tools:layout="@layout/fragment_people_details">
         <argument
             android:name="people"
-            app:argType="com.sw.sw_api_kotlin_project.network.model.People" />
+            app:argType="com.sw.sw_api_kotlin_project.data.network.model.People" />
     </fragment>
 
     <fragment
@@ -58,7 +58,7 @@
         tools:layout="@layout/fragment_film_details">
         <argument
             android:name="film"
-            app:argType="com.sw.sw_api_kotlin_project.network.model.Film" />
+            app:argType="com.sw.sw_api_kotlin_project.data.network.model.Film" />
     </fragment>
 
     <fragment
@@ -78,7 +78,7 @@
         tools:layout="@layout/fragment_planet_details">
         <argument
             android:name="planet"
-            app:argType="com.sw.sw_api_kotlin_project.network.model.Planet" />
+            app:argType="com.sw.sw_api_kotlin_project.data.network.model.Planet" />
     </fragment>
 
     <fragment
