+++ SW_API_Kotlin_Project/data/src/main/java/com/sw/sw_api_kotlin_project/data/network/model/Film.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.network.model
+package com.sw.sw_api_kotlin_project.data.network.model
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
