+++ SW_API_Kotlin_Project/data/src/main/java/com/sw/sw_api_kotlin_project/data/network/model/Starships.kt
@@ -1,11 +1,16 @@
-package com.sw.sw_api_kotlin_project.network.model
+package com.sw.sw_api_kotlin_project.data.network.model
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class Vehicles(
+data class Starships(
+    /**
+     * 近代銀河光時間
+     */
+    @SerializedName("MGLT")
+    val mglt: String,
     @SerializedName("cargo_capacity")
     val cargoCapacity: String,
     val consumables: String,
@@ -15,6 +20,8 @@ data class Vehicles(
     val crew: String,
     val edited: String,
     val films: List<String>,
+    @SerializedName("hyperdrive_rating")
+    val hyperDriveRating: String,
     val length: String,
     val manufacturer: String,
     @SerializedName("max_atmosphering_speed")
@@ -23,6 +30,6 @@ data class Vehicles(
     val name: String,
     val passengers: String,
     val pilots: List<String>,
-    val url: String,
-    val vehicle_class: String
+    val starship_class: String,
+    val url: String
 ) : Parcelable
